Refactor service into routes
