+++ docs/nav.html
@@ -19,14 +19,14 @@
 				<a href="/v{{apiversion}}/docs/contributing">Contributing</a>
 				{{#ifPrefix pageName "contributing"}}
 				<ul class="o-techdocs-nav o-techdocs-nav--sub">
-					<li{{sectionHighlight pageName "contributing/authoring-polyfills"}}>
-						<a href="/v{{apiversion}}/docs/contributing/authoring-polyfills">Authoring polyfills</a>
+					<li{{sectionHighlight ../pageName "contributing/authoring-polyfills"}}>
+						<a href="/v{{../apiversion}}/docs/contributing/authoring-polyfills">Authoring polyfills</a>
 					</li>
-					<li{{sectionHighlight pageName "contributing/testing"}}>
-						<a href="/v{{apiversion}}/docs/contributing/testing">Testing</a>
+					<li{{sectionHighlight ../pageName "contributing/testing"}}>
+						<a href="/v{{../apiversion}}/docs/contributing/testing">Testing</a>
 					</li>
-					<li{{sectionHighlight pageName "contributing/common-scenarios"}}>
-						<a href="/v{{apiversion}}/docs/contributing/common-scenarios">Common scenarios</a>
+					<li{{sectionHighlight ../pageName "contributing/common-scenarios"}}>
+						<a href="/v{{../apiversion}}/docs/contributing/common-scenarios">Common scenarios</a>
 					</li>
 				</ul>
 				{{/ifPrefix}}
