+++ lib/UA.js
@@ -138,7 +138,7 @@ function UA(uaString) {
 
 			// Map to different family with per-version mapping
 			} else if (typeof aliases[this.ua.family] === 'object') {
-				for (let semverExpr in aliases[this.ua.family]) {
+				for (let semverExpr in aliases[this.ua.family]) {   // eslint-disable-line prefer-const
 					if (this.ua.satisfies(semverExpr) && Array.isArray(aliases[this.ua.family][semverExpr])) {
 						const a = aliases[this.ua.family][semverExpr];
 						this.ua = new useragent.Agent(a[0], a[1], (a[2] || 0), (a[3] || 0));
