+++ lib/index.js
@@ -91,7 +91,7 @@ function getPolyfills(options) {
 			}
 		});
 		return features;
-	}
+	};
 
 	return Promise.resolve(options.features)
 		.then(resolveAliases)
@@ -116,8 +116,9 @@ function getPolyfillString(options) {
 			if (options.minify) {
 				explainerComment.push('Rerun without minification (remove `.min` from URL path) for verbose metadata');
 			} else {
+				const versionStr = process.env.NODE_ENV === 'production' ? 'v'+appVersion : 'DEVELOPMENT MODE - for live use set NODE_ENV to \'production\'';
 				explainerComment.push(
-					'Polyfill service v' + appVersion,
+					'Polyfill service ' + versionStr,
 					'For detailed credits and licence information see http://github.com/financial-times/polyfill-service.',
 					'',
 					'UA detected: ' + uaDebugName,
@@ -222,7 +223,7 @@ function getPolyfillString(options) {
 				}
 
 				if ('all' in options.features) {
-					let warnText = 'Using the `all` alias with polyfill.io is a very bad idea. In a future version of the service, `all` will deliver the same behaviour as `default`, so we recommend using default instead.';
+					const warnText = 'Using the `all` alias with polyfill.io is a very bad idea. In a future version of the service, `all` will deliver the same behaviour as `default`, so we recommend using default instead.';
 					explainerComment.push('', warnText);
 					builtPolyfillString += "\nconsole.log('"+warnText+"');\n";
 				}
